Added contexts to filters, added FilterWrapper
The FilterWrapper is meant to enforce pipeline deadlines
once implemented and perform common tasks across all Filters
regardless of if they are part of the core or not. Eventually
this will allow for Logstash like functionality where if a filter
is successful, fields or tags can be removed/added.

If statements will be implemented as a superset of a FilterWrapper.