Give filters matchFunc, removed end, split mutate, grok ignore_field
Filters are now given a function to call if the filter matches an
event. The func is give by the FilterWrapper and it's up to the Filter
to call it for each event matched.

The End filter isn't needed anymore so it's removed.

The Mutate filter is now the remove_field filter. Each "mutate"
filter will be a separate filter. These filters will not call
the matchFunc as these mutate filters are meant to do a specific
thing and nothing more.

The grok filter has a new option named ignore_field that will not
add "_grokparsefailure" to the event tags if the field doesn't exist.