Provide base struct types for input,output,codec
The idea being that internal, mundane tasks can be
handled by the core simply by embeding the base struct
into a plugin struct. For example codec creation
can be handled by the base struct and be disabled
by overriding the LoadConfig method on a plugin that
doesn't need/want it.